Changes Made 	1. Renamed all Files from js to jsx extension 	2. Added React Router functionality 	3. Create Components for each Side Nav Menu 	4. Setting Breadcrumbps moved to each SideNav Component 	5. Setting ActiveSideNav moved to each SideNav Component 	6. Setting ActivePage moved to each Page Component 	7. Changed the tab icon to the app Icon 	8. Added Icon Provider for Non-Social Side Nave Menu 	9. Changed const pages to a array of JSON with key title and link.
