--- echoo/src/Resources/image/Behind_PXL_Logo_SVG_Data.js
@@ -1,98 +1,0 @@
-const Behind_Pixl_logo= () => (
-    <svg 
-        version="1.0" xmlns="http://www.w3.org/2000/svg"
-        width="20.000000pt" height="20.000000pt" viewBox="0 0 2000.000000 2000.000000"
-        preserveAspectRatio="xMidYMid meet"
-    >
-        <g 
-            transform="translate(0.000000,2000.000000) scale(0.100000,-0.100000)"
-            fill="#FFFFFF" stroke="none"
-        >
-            <path d="M9310 15098 c-1 -754 -10 -1672 -17 -1678 -4 -3 -3 3 0 13 6 15 4 16
-            -13 7 -10 -5 -22 -15 -25 -20 -4 -6 -10 -8 -15 -5 -5 3 -11 1 -15 -5 -4 -6
-            -11 -7 -17 -4 -6 4 -8 2 -5 -3 10 -16 -18 -24 -33 -10 -10 9 -11 9 -6 0 5 -8
-            1 -13 -9 -13 -10 0 -14 5 -9 13 6 10 4 10 -7 0 -16 -16 -33 -17 -23 -1 5 8 3
-            8 -7 2 -8 -5 -1402 -10 -3369 -10 -1845 -1 -3453 -2 -3572 -3 l-218 -1 0 -23
-            c0 -17 31 -54 108 -127 59 -57 402 -398 763 -757 l657 -653 4048 -3 4049 -3
-            200 -21 c251 -27 522 -70 675 -108 137 -34 306 -78 317 -84 15 -8 -94 -138
-            -186 -223 -195 -181 -479 -328 -706 -366 -68 -12 -153 -11 -174 2 -10 6 -12 5
-            -7 -2 4 -7 2 -12 -3 -12 -6 0 -11 4 -11 8 0 5 -7 9 -15 9 -8 0 -15 -5 -15 -11
-            0 -5 -7 -1 -16 10 -8 10 -13 13 -9 7 3 -7 2 -13 -2 -13 -5 0 -15 6 -22 13 -11
-            10 -13 10 -7 0 4 -7 3 -13 -3 -13 -5 0 -12 6 -14 13 -4 10 -6 10 -6 0 -1 -7
-            -6 -13 -12 -13 -5 0 -7 6 -3 13 4 6 4 9 -1 5 -4 -4 -15 -2 -23 5 -13 9 -14 9
-            -8 0 4 -7 3 -13 -1 -13 -5 0 -15 6 -22 13 -11 10 -13 10 -7 0 4 -8 1 -13 -8
-            -13 -9 0 -16 6 -16 13 0 8 -4 7 -10 -3 -5 -8 -10 -10 -10 -3 0 6 -8 10 -17 8
-            -10 -1 -24 3 -32 9 -11 8 -13 8 -7 -1 8 -14 -6 -18 -18 -6 -10 10 -87 12 -81
-            2 3 -4 0 -11 -6 -15 -7 -4 -9 -3 -6 3 4 6 -1 17 -10 24 -9 8 -14 9 -10 3 4 -7
-            0 -14 -8 -17 -10 -4 -12 -1 -8 11 3 9 3 14 -2 10 -4 -4 -15 -2 -23 5 -10 7
-            -13 8 -9 1 4 -7 -2 -14 -13 -17 -11 -3 -23 0 -27 5 -3 6 -1 8 5 4 6 -3 13 -3
-            16 2 2 4 -16 8 -42 7 -26 -1 -50 4 -54 9 -5 7 -8 7 -8 -1 0 -7 -5 -13 -11 -13
-            -5 0 -7 6 -3 13 4 6 4 9 -1 5 -4 -4 -15 -2 -23 5 -13 9 -14 9 -8 0 5 -9 0 -13
-            -13 -13 -12 0 -21 6 -21 13 0 8 -4 7 -10 -3 -9 -13 -10 -13 -10 0 0 13 -1 13
-            -10 0 -9 -14 -14 -7 -11 13 1 5 -3 6 -9 2 -6 -3 -8 -10 -5 -15 3 -5 0 -12 -6
-            -16 -7 -4 -9 -3 -6 2 3 6 2 16 -4 24 -7 13 -9 13 -9 1 0 -11 -4 -9 -15 5 -15
-            19 -15 19 -8 -1 l6 -20 -23 20 c-13 11 -20 14 -16 8 4 -7 2 -13 -3 -13 -6 0
-            -11 -16 -11 -35 l0 -35 -775 0 c-613 0 -775 3 -776 13 -1 6 -1 17 0 22 0 6
-            -15 11 -34 11 -19 0 -35 -4 -35 -8 0 -4 -5 -8 -11 -8 -5 0 -7 6 -3 13 6 9 4 9
-            -7 0 -17 -17 -32 -17 -23 0 6 9 4 9 -7 1 -8 -5 -21 -10 -29 -9 -8 0 -21 -6
-            -28 -15 -11 -13 -11 -13 -7 1 4 10 0 16 -9 16 -8 0 -13 -4 -10 -9 3 -4 1 -8
-            -5 -8 -5 0 -12 6 -14 13 -4 9 -6 8 -6 -2 -1 -10 -4 -11 -15 -2 -9 8 -16 8 -20
-            2 -4 -6 -14 -7 -23 -4 -12 5 -14 3 -8 -8 7 -12 6 -12 -7 -1 -12 9 -18 10 -23
-            2 -5 -8 -11 -8 -21 1 -11 9 -14 8 -14 -2 0 -11 -2 -11 -10 1 -6 10 -10 11 -10
-            3 0 -7 -5 -13 -11 -13 -5 0 -7 6 -3 13 6 9 5 9 -7 0 -11 -9 -19 -9 -26 -2 -8
-            8 -13 7 -18 -1 -5 -8 -11 -7 -23 2 -12 10 -14 10 -8 1 4 -7 3 -13 -3 -13 -5 0
-            -13 5 -16 10 -5 8 -11 8 -19 1 -7 -6 -16 -8 -19 -4 -4 3 -7 0 -7 -8 0 -11 -2
-            -11 -10 1 -9 13 -10 13 -10 0 0 -13 -1 -13 -10 0 -9 13 -10 13 -10 0 0 -13 -1
-            -13 -10 0 -6 10 -10 11 -10 3 0 -7 -5 -13 -11 -13 -5 0 -8 4 -5 8 3 5 -1 9 -9
-            9 -8 0 -12 -4 -9 -9 3 -4 -2 -8 -11 -8 -10 0 -13 5 -9 13 6 9 4 9 -8 -1 -11
-            -9 -18 -10 -23 -2 -3 6 -12 7 -18 3 -20 -12 -97 -11 -90 1 5 8 3 8 -8 0 -10
-            -8 -605 -11 -2107 -11 -1151 -1 -2096 -4 -2100 -7 -3 -4 349 -360 784 -791
-            l789 -784 2985 -1 c2625 0 2993 -2 3050 -15 178 -42 344 -114 500 -216 113
-            -74 301 -258 379 -369 191 -273 285 -573 284 -910 -1 -566 -314 -1089 -803
-            -1344 -131 -68 -298 -127 -367 -129 -18 -1 -33 -5 -33 -10 0 -5 -8 -3 -17 4
-            -10 8 -14 9 -10 2 6 -9 -87 -12 -393 -12 -306 0 -399 3 -393 12 4 7 0 6 -10
-            -2 -11 -8 -17 -9 -17 -3 0 7 -7 9 -17 5 -12 -5 -14 -3 -8 8 7 12 6 12 -7 1
-            -14 -11 -20 -9 -39 10 -12 13 -19 18 -16 13 3 -6 -2 -11 -12 -11 -10 0 -18 6
-            -18 13 0 7 5 9 11 6 6 -4 7 -2 3 3 -5 5 -14 9 -20 9 -7 0 -12 5 -12 12 0 6 -4
-            15 -9 20 -6 5 -6 2 -1 -7 8 -13 6 -13 -8 -1 -16 12 -17 94 -19 1095 l-3 1081
-            -774 3 c-490 1 -778 6 -786 12 -6 5 -15 7 -19 3 -8 -9 -572 -13 -1778 -13
-            -478 0 -868 -4 -868 -8 0 -4 356 -361 792 -794 l793 -787 543 -1 542 0 0
-            -2330 0 -2330 1255 0 c830 0 1255 3 1255 10 0 6 15 10 33 10 42 0 317 43 482
-            76 1027 202 1983 658 2777 1326 169 142 542 515 678 678 560 672 948 1445
-            1129 2247 91 404 122 665 128 1083 5 334 -5 505 -48 805 -138 970 -552 1886
-            -1210 2675 -192 231 -547 579 -779 765 -47 37 -84 72 -83 76 2 5 -3 8 -10 6
-            -6 -1 -11 4 -10 10 2 7 -3 12 -10 10 -6 -1 -11 4 -10 10 2 7 -3 12 -10 10 -6
-            -1 -11 4 -9 11 1 8 -1 11 -5 9 -11 -7 -34 19 -25 28 4 4 0 5 -8 3 -8 -2 -14 3
-            -12 10 1 8 -2 11 -7 8 -12 -8 -42 34 -34 47 3 6 1 7 -4 4 -22 -14 -31 21 -32
-            123 -3 196 -30 414 -73 584 -284 1134 -1080 1987 -2203 2361 -150 49 -400 112
-            -548 136 -286 47 -339 49 -1499 49 l-1098 0 0 -532z m2651 -1113 c353 -92 647
-            -305 855 -616 66 -100 71 -109 53 -109 -5 0 -7 6 -3 13 6 10 4 10 -7 0 -15
-            -15 -33 -18 -25 -5 3 5 -1 9 -9 9 -8 0 -12 -4 -9 -9 9 -15 -19 -8 -39 10 -15
-            12 -18 13 -12 2 7 -12 6 -13 -6 -4 -8 6 -24 10 -36 8 -13 -2 -23 0 -23 5 0 5
-            -8 8 -17 6 -10 -1 -22 0 -28 3 -5 3 -16 7 -25 8 -8 1 -31 5 -50 9 -19 4 -48 9
-            -65 11 -71 11 -80 12 -80 14 0 2 -11 4 -85 14 -19 2 -35 5 -35 6 0 1 -20 3
-            -45 5 -25 1 -88 6 -140 11 -52 4 -300 8 -552 8 -303 1 -454 4 -451 11 3 5 0 4
-            -8 -2 -7 -7 -17 -13 -22 -13 -4 0 -5 6 -1 13 5 8 2 8 -9 -2 -14 -11 -20 -11
-            -32 -1 -8 7 -15 9 -15 5 0 -4 -8 -1 -17 6 -11 10 -14 10 -9 2 4 -7 2 -13 -3
-            -13 -6 0 -11 5 -11 12 0 6 -3 8 -7 5 -3 -4 -13 -1 -21 6 -13 10 -15 9 -8 -3 4
-            -8 0 -6 -8 4 -9 11 -16 14 -16 8 0 -7 -5 -12 -11 -12 -5 0 -7 6 -3 13 5 7 2 7
-            -10 -2 -15 -12 -17 -12 -14 0 2 8 -3 13 -9 12 -7 -2 -13 2 -13 8 0 5 -4 7 -10
-            4 -7 -4 -10 92 -10 279 l0 286 504 0 c504 0 504 0 597 -25z m2413 -3378 c375
-            -409 604 -758 799 -1222 346 -819 375 -1759 82 -2630 -95 -281 -260 -613 -438
-            -880 -66 -100 -226 -317 -246 -335 -3 -3 -33 -36 -66 -75 -186 -217 -413 -425
-            -639 -584 -341 -241 -779 -446 -1204 -565 -120 -34 -335 -84 -391 -91 -23 -3
-            -46 -7 -50 -9 -4 -3 -74 -14 -157 -26 -139 -20 -177 -24 -344 -37 -36 -3 -198
-            -6 -360 -8 -261 -2 -297 -1 -314 14 -11 9 -16 10 -12 4 4 -7 2 -13 -3 -13 -6
-            0 -11 5 -11 11 0 5 -4 8 -9 5 -5 -3 -14 1 -21 9 -11 14 -15 14 -34 -2 -14 -13
-            -17 -13 -9 -2 11 13 6 21 -25 50 -32 30 -34 30 -9 3 15 -16 27 -34 27 -39 0
-            -5 -7 1 -16 13 -8 12 -20 20 -25 16 -5 -3 -8 -2 -7 3 1 4 -3 18 -10 29 -8 16
-            -11 100 -9 300 l2 279 40 -2 c201 -9 588 -4 720 11 754 83 1361 325 1903 759
-            553 443 943 1065 1126 1795 83 330 85 749 5 1133 -74 361 -227 714 -462 1069
-            -113 171 -193 266 -387 460 l-175 175 170 170 c94 94 212 223 264 287 l94 118
-            58 -51 c32 -28 96 -92 143 -142z"/>
-        </g>
-    </svg>
-); 
-
-
-export default Behind_Pixl_logo;
