+++ echoo/src/index.js
@@ -1,17 +1,21 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import App from './App';
 import reportWebVitals from './reportWebVitals';
+
 // Redux imports
 import { Provider } from 'react-redux';
 import {store} from './Redux/Store'
+
+// Router Imports
+import MyRouter from "./Router/MyRouter"
+import { RouterProvider } from "react-router-dom";
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <Provider store={store}>
     <React.StrictMode>
-      <App />
+      <RouterProvider router={MyRouter} />
     </React.StrictMode>
   </Provider>
 );
